Test nullable body (#2648)
* Test nullable body

Test the 2 possible variants for nullable body:
- type: ['object', 'null']
- type: 'object', nullable: true

See: https://github.com/fastify/fastify/issues/2618.

* Improve POST request test

As noticed by @jsumners, `fastify.inject` will pass a null body in
the POST request in the abscence of a body, which may lead to false
positives. What we want is a POST request that pass an empty body,
not a body with "null", which is a valid JSON.

By using sget, it makes a POST request which pass an empty body.

See https://github.com/fastify/fastify/pull/2648#issuecomment-716505296.